requirements.txt
rumps
sounddevice
numpy
requests
websockets
pyobjc-framework-Quartz
pyobjc-framework-HIServices
pyobjc-framework-AppKit
